+++ modules/apps/commerce/commerce-product-definitions-web/src/main/java/com/liferay/commerce/product/definitions/web/internal/frontend/taglib/clay/data/set/filter/ProductTypeClayTableDataSetFilter.java
@@ -12,13 +12,21 @@
  * details.
  */
 
-package com.liferay.commerce.product.definitions.web.internal.frontend;
+package com.liferay.commerce.product.definitions.web.internal.frontend.taglib.clay.data.set.filter;
 
 import com.liferay.commerce.product.definitions.web.internal.frontend.constants.CommerceProductDataSetConstants;
-import com.liferay.frontend.taglib.clay.data.set.filter.BaseAutocompleteClayDataSetFilter;
+import com.liferay.commerce.product.type.CPType;
+import com.liferay.commerce.product.type.CPTypeServicesTracker;
+import com.liferay.frontend.taglib.clay.data.set.filter.BaseRadioClayDataSetFilter;
 import com.liferay.frontend.taglib.clay.data.set.filter.ClayDataSetFilter;
+import com.liferay.frontend.taglib.clay.data.set.filter.RadioClayDataSetFilterItem;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marco Leo
@@ -28,37 +36,36 @@ import org.osgi.service.component.annotations.Component;
 	property = "clay.data.set.display.name=" + CommerceProductDataSetConstants.COMMERCE_DATA_SET_KEY_PRODUCT_DEFINITIONS,
 	service = ClayDataSetFilter.class
 )
-public class CommerceCatalogClayTableDataSetFilter
-	extends BaseAutocompleteClayDataSetFilter {
-
-	@Override
-	public String getAPIURL() {
-		return "/o/headless-commerce-admin-catalog/v1.0/catalogs?sort=name:asc";
-	}
+public class ProductTypeClayTableDataSetFilter
+	extends BaseRadioClayDataSetFilter {
 
 	@Override
 	public String getId() {
-		return "catalogId";
-	}
-
-	@Override
-	public String getItemKey() {
-		return "id";
-	}
-
-	@Override
-	public String getItemLabel() {
-		return "name";
+		return "productType";
 	}
 
 	@Override
 	public String getLabel() {
-		return "catalog";
+		return "product-type";
 	}
 
 	@Override
-	public boolean isMultipleSelection() {
-		return false;
+	public List<RadioClayDataSetFilterItem> getRadioClayDataSetFilterItems(
+		Locale locale) {
+
+		List<RadioClayDataSetFilterItem> radioClayDataSetFilterItems =
+			new ArrayList<>();
+
+		for (CPType cpType : _cpTypeServicesTracker.getCPTypes()) {
+			radioClayDataSetFilterItems.add(
+				new RadioClayDataSetFilterItem(
+					cpType.getLabel(locale), cpType.getName()));
+		}
+
+		return radioClayDataSetFilterItems;
 	}
+
+	@Reference
+	private CPTypeServicesTracker _cpTypeServicesTracker;
 
 }
